Cache Artifactory API responses
When requests to the Artifactory API time out, we get incomplete data in
the metrics output. This change circumvents this by caching the last
successful request to Artifactory for each HTTP method/path/body.

This is intended as a workaround for large Artifactory clusters with
long response times, where increasing the ARTI_TIMEOUT would just lead
to scrape timeouts from Prometheus.

Configuration:
* USE_CACHE (default: false) Whether to activate the feature. If set to
  false, caching will be skipped.
* CACHE_TIMEOUT (default: 30s) Limit for requests before cached
  responses are served. Independent of the ARTI_TIMEOUT, so
  CACHE_TIMEOUT can be set to less than the scrape interval, while
  ARTI_TIMEOUT can be set to a higher value to accomodate slow
  artifactory responses.
  Should be set to a lower value than ARTI_TIMEOUT if the cache is in
  use.
* CACHE_TTL (default: 5m) Maximum time for the cache entry to be valid.
  After expiry, cache entries will be ignored.
  Useful so that requests that keep failing don't show up as successful.